requirements.txt
mistune
wheel>=0.24.0
Django>=1.9
xxhash==0.4.3
celery[redis]
django-crispy-forms>=1.6.0
six
pillow
djangorestframework
